Abstract RoomManager and Multiplayer
